\name{BQObject-class}
\docType{class}
\alias{BQObject-class}

\title{Class "BQObject"}
\description{
    Parent class to other BQ objects Driver, Connection, and Result.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
  \describe{
    \item{\code{Id}:}{Unique object id for be BQObject}
  }
}
\section{Extends}{
Class \code{"\linkS4class{DBIObject}"}, directly.
Class \code{"\linkS4class{dbObjectId}"}, directly.
}
\section{Methods}{
No methods defined with class "BQObject" in the signature.
}
\seealso{

\code{\link[DBI]{DBIObject-class}}
\code{\link{BQDriver-class}}
\code{\link{BQConnection-class}}
\code{\link{BQResult-class}}

}
\keyword{classes}
