\name{summary}
\alias{summary,BQResult-method}
\alias{summary,BQConnection-method}
\alias{summary,BQDriver-method}
\title{Print a short summary of a bigquery object}
\description{
    Prints in human readable form information about a BQObject.
}
\usage{
    summary( obj, verbose=FALSE )
}
\seealso{
    \code{\link{dbGetInfo}}
}
